refactor: improve code quality and structure
