--- src/components/pages/free-tours/DestinationFilter/icons/ChevronRightIcon.tsx
@@ -1,25 +1,0 @@
-type TChevronRightIconProps = {
-  className?: string
-}
-
-const ChevronRightIcon = ({ className = '' }: TChevronRightIconProps) => {
-  return (
-    <svg
-      width='16'
-      height='16'
-      viewBox='0 0 16 16'
-      fill='none'
-      xmlns='http://www.w3.org/2000/svg'
-      className={className}
-    >
-      <path
-        d='M6 12L10 8L6 4'
-        stroke='currentColor'
-        strokeLinecap='round'
-        strokeLinejoin='round'
-      />
-    </svg>
-  )
-}
-
-export default ChevronRightIcon
