+++ src/components/shared/icons/common/CheckIcon.tsx
