+++ src/components/shared/icons/common/ClearIcon.tsx
