+++ src/components/shared/icons/common/CloseIcon.tsx
